Fixes i18next/i18next-fluent#1
Use the i18next `preload` option to load languages on top of the ones
detected by i18next after it is initialized. Fixes an issue server-side
where the language detection only runs during a request and not during
the initialization step, resulting in the language loading throwing an
error.

The fix was tested server-side in combination with 18next-node-fs-backend
and i18next-express-middleware. It was also tested client-side in
combination with i18next-fluent-backend and
i18next-browser-languagedetector.